moved user class into user module in models and user endpoints into user service under services
