+++ src/services/user_service.py
@@ -1,0 +1,18 @@
+from fastapi import HTTPException
+from sqlmodel import Session
+from src.main import User  # Adjust import as needed
+
+def create_user_service(user: User, session: Session):
+    session.add(user)
+    session.commit()
+    session.refresh(user)
+    return {
+        "message": f"User {user.name} created successfully!",
+        "data": user
+    }
+
+def read_user_service(user_id: int, session: Session):
+    user = session.get(User, user_id)
+    if not user:
+        raise HTTPException(status_code=404, detail="User not found")
+    return user
